Add both demos
